+++ Sources/Shared/Toolkit/Benchmark.swift
@@ -7,14 +7,14 @@
 import Foundation
 
 /// Measures the execution time of `block`.
-public func measure<T>(_ tag: String, _ block: () -> T) -> T {
+package func benchmark<T>(_ tag: String, _ block: () -> T) -> T {
     let startPoint = Date()
     defer { print("⏱️ Measure \(tag): \(Date().timeIntervalSince(startPoint)) seconds") }
     return block()
 }
 
 /// Measures the execution time of `block`.
-public func measure<T>(_ tag: String, _ block: () async -> T) async -> T {
+package func benchmark<T>(_ tag: String, _ block: () async -> T) async -> T {
     let startPoint = Date()
     defer { print("⏱️ Measure \(tag): \(Date().timeIntervalSince(startPoint)) seconds") }
     return await block()