+++ Sources/Shared/Toolkit/Data/Streamable.swift
@@ -34,6 +34,7 @@ public extension Streamable {
     /// - Parameters:
     ///   - consume: Callback called for each chunk of data received. Callers
     ///     are responsible to accumulate the data if needed.
+    // FIXME: Task cancellation
     func stream(consume: @escaping (Data) -> Void) async -> ReadResult<Void> {
         await stream(range: nil, consume: consume)
     }
@@ -85,5 +86,64 @@ public extension Streamable {
     @available(*, deprecated, message: "Use `read().asJSONObject()` instead")
     func readAsJSONObject(options: JSONSerialization.ReadingOptions = []) async -> ReadResult<[String: Any]> {
         await readAsJSON()
+    }
+}
+
+package extension Streamable {
+    /// Reads the whole bytes while monitoring available memory to avoid OOM
+    /// crashes.
+    func readMonitoringMemory() async throws(ReadError) -> Data {
+        let estimated = await estimatedLength().getOrNil() ?? nil
+
+        guard !Task.isCancelled else {
+            throw .cancelled
+        }
+
+        // `availableMemory` will be 0 on the Simulator.
+        let availableMemory = os_proc_available_memory()
+        if availableMemory > 0, let length = estimated, length > availableMemory {
+            throw .outOfMemory(nil)
+        }
+
+        var data = Data()
+        if let estimated, estimated <= UInt64(Int.max) {
+            data.reserveCapacity(Int(estimated))
+        }
+
+        var error: ReadError? = nil {
+            didSet {
+                data = Data()
+            }
+        }
+
+        let streamResult = await stream { chunk in
+            guard error == nil else {
+                return
+            }
+
+            guard !Task.isCancelled else {
+                error = .cancelled
+                return
+            }
+
+            let availableMemory = os_proc_available_memory()
+            guard availableMemory == 0 || data.count + chunk.count <= availableMemory else {
+                error = .outOfMemory(nil)
+                return
+            }
+
+            data.append(chunk)
+        }
+
+        if let error {
+            throw error
+        }
+
+        switch streamResult {
+        case .success:
+            return data
+        case let .failure(error):
+            throw error
+        }
     }
 }